+++ cosipy/__init__.py
@@ -5,6 +5,7 @@ from .response import DetectorResponse
 from .data_io import DataIO
 from .data_io import UnBinnedData
 from .data_io import BinnedData
+from .data_io import ReadTraTest
 from .make_plots import MakePlots
 
 from .threeml import COSILike
