+++ cosipy/data_io/__init__.py
@@ -1,3 +1,4 @@
 from .DataIO import DataIO
 from .UnBinnedData import UnBinnedData
 from .BinnedData import BinnedData
+from .ReadTraTest import ReadTraTest
